pyproject.toml
[project]
    name = "pydevice2mqtt"
    description = '''
    This project provides a simplified control of devices via MQTT.
    Supports the Hasso auto configuration protocol
    '''
    license = {file = "LICENSE"}
    version = "0.1.1"
    readme = "README.md"
    requires-python = ">=3.9"
    authors = [{name="HighImp", email="AnyHighZ@gmail.com"}]
    dependencies = [
        "paho-mqtt",
        "pyyaml"
    ]
    keywords = ["hassio","mqtt","homeassistent","homeautomation"]
    repository = "https://github.com/HighImp/pydevice2mqtt"

[project.scripts]
    flit = "flit:main"


[project.optional-dependencies]
    test = [
        "pylint ~=2.14.0",
        "pytest-cov ~=4.0.0",
        "pytest-mock ~=3.10.0"
    ]

[build-system]
    build-backend = "flit_core.buildapi"
    requires = ["flit_core >=3.8.0,<4"]

[tool.pytest.ini_options]
    addopts = "--cov --cov-report html --cov-report term-missing --cov-fail-under 60"

[tool.coverage.run]
    source = ["src"]